Move AprilTag code to 'perception' subdirectory
